backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
anthropic>=0.42.0
pypdf>=5.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
pydantic>=2.10.0
aiosqlite>=0.20.0
